add `nn.functional.upsample_nearest` to fin_grain and modify the file structure
