+++ test/xpu/fin_grain/test_ops_xpu.py
@@ -134,6 +134,7 @@ _xpu_computation_op_list = [
     "nn.functional.embedding",
     "nn.functional.unfold",
     "nn.functional.interpolate",
+    "nn.functional.upsample_nearest",
     # "nn.functional.nll_loss", # Lack of XPU implementation of aten::nll_loss2d_forward. Will retrieve the case, only if the op is implemented.
     "sigmoid",
     "sgn",
